Remove leeway.
Leeway is not necessary in the relay pledge. The server can know every
stored message prior to the find's block. Even if the find goes through
the on chain queue, the server can know everything that might be forced
or let through earlier. It just needs to ensure it doesn't store any
completed blocks.