fix(RelayPledge): request messages before a point
The initial implementation allowed you to request a message after a
given block number and reference message. That meant that finding the
latest message by a user would require binary search, majorly increasing
the time-to-service the server could impose.
Now a user can ask for the first message before the current block in a
single request.