feat(RelayPledge): prefix matching
The user can ask the server for a message with a given prefix. This
enables messages of different types to be stores. For example, we can
store setting on the server, or direct messages, or public posts, with
varying levels of encryption, and request the latest state for any kind.