refactor: don't enforce strict Pledge interface
Having an isBroken method that covered different types of pledges
introduced repeat slashing bugs and ugly code. In the case of the
LivelinessPledge, we needed to check that the supplied receipts included
nothing but the message or message hash, otherwise an attacker could
alter some other aspect of the calldata and a slash the contract many
times for only one genuine infraction by the server.

This format will require extra work in the bond contract to call
isBroken methods with different parameters.